fix: playing time bar
Sometimes, the bar isn't full when playback ends.
From my understandings, this should be an VLC issue: when playback ends, the timing may not be the duration of the song (like 3:27 / 3:28).
The only way to fix this is to force round / ceiling of bar.Fraction. I've left some tries I made and let's check out...